refactor(knowledge): relocate findUnansweredQuestionLogic into query (#209)
