+++ src/architecture/knowledge/synthesis/evidenceMap.ts
@@ -2,7 +2,7 @@ import type { KnowledgeModel } from "../model/KnowledgeModel";
 import type { Source } from "../model/Idea";
 import { incomingRelations, outgoingRelations } from "../query/queries";
 import { findContradictions } from "architecture/knowledge/query/findContradictionLogic";
-import { findUnansweredQuestions } from "actions/findUnansweredQuestion/findUnansweredQuestionLogic";
+import { findUnansweredQuestions } from "architecture/knowledge/query/findUnansweredQuestionLogic";
 
 /** The relation that backs a position (#147). */
 const SUPPORTS = "supports";
